implement getSimilarSessionsDistances function and integrate it into MealTemplate
