+++ src/models/mealTemplate.ts
@@ -1,4 +1,8 @@
 import { getBasalSensitivity } from "../lib/basal";
+import {
+  getSimilarSessionsDistances,
+  type SessionAndScore,
+} from "../lib/helpers/getSimilarSessionsDistances";
 import { sessionsWeightedAverage } from "../lib/templateHelpers";
 import { timeOfDayOffset } from "../lib/timing";
 import { clamp, MathUtil } from "../lib/util";
@@ -9,13 +13,10 @@ import { PreferencesStore } from "../storage/preferencesStore";
 import { PrivateStore } from "../storage/privateStore";
 import Insulin from "./events/insulin";
 import type Meal from "./events/meal";
-import type Food from "./food";
 import Session from "./session";
 import Subscribable from "./subscribable";
 import type { Template } from "./types/interfaces";
 import type { Deserializer, Serializer } from "./types/types";
-
-type SessionAndScore = [Session, number];
 
 export default class MealTemplate extends Subscribable implements Template {
   sessions: Session[] = [];
@@ -310,47 +311,7 @@ export default class MealTemplate extends Subscribable implements Template {
       (s) => s.meals.length === 1,
     );
     if (sessions.length === 0) return null;
-
-    function toFoodMap(foods: Food[]): Map<string, number> {
-      const map = new Map<string, number>();
-      for (const food of foods) {
-        map.set(food.name, (map.get(food.name) ?? 0) + food.amount);
-      }
-      return map;
-    }
-
-    const targetMap = toFoodMap(meal.summedFoods);
-    const sessionMaps = sessions.map((s) => toFoodMap(s.firstMeal.summedFoods));
-
-    const allFoodNames = new Set<string>();
-    for (const map of sessionMaps) {
-      for (const name of map.keys()) allFoodNames.add(name);
-    }
-    for (const name of targetMap.keys()) allFoodNames.add(name);
-
-    const foodScales = new Map<string, number>();
-    for (const name of allFoodNames) {
-      const amounts = sessionMaps.map((map) => map.get(name) ?? 0);
-      const mad = MathUtil.meanAbsoluteDeviation(amounts);
-      foodScales.set(name, Math.max(mad, 2e-3));
-    }
-
-    const sessionsDistances: SessionAndScore[] = sessions.map((session, i) => {
-      const sessionMap = sessionMaps[i];
-      let dist = 0;
-
-      for (const name of allFoodNames) {
-        const a = sessionMap.get(name) ?? 0;
-        const b = targetMap.get(name) ?? 0;
-        const scale = foodScales.get(name) ?? 1;
-        dist += ((a - b) / scale) ** 2;
-      }
-
-      return [session, dist];
-    });
-
-    sessionsDistances.sort((a, b) => a[1] - b[1]); // Sort from smallest distance to largest
-    return sessionsDistances;
+    return getSimilarSessionsDistances(meal, sessions);
   }
   getBaseSession(meal: Meal): Session | null {
     let sessionsDistances = this.getSimilarSessionsDistances(meal);
